Working on rails 3.0
The main difference is that 3.0 still used a Hash subclass as the
FlashHash, ugh. This also breaks the marshal/unmarshal between 3.0 and
3.1. But this was a known issue when upgrading from rails 3.0 to 3.1.
This makes this gem a good option for a cleaner upgrade path.

Also, Flash::KEY is first defined in 3.1, but it's nice to have a frozen
string constant so backported.